services/proxy_resolver: Let ProxyResolvers outlive ProxyResolverFactory
Make ProxyResolvers hold onto ServiceContextRefs. This will allow
the ServiceManager to control the lifetime of the proxy_resolver
service, shutting it down when there are no ProxyResolvers, so
ChromeMojoProxyResolverFactory will no longer need to micromanage its
lifetime.

We'll still need a ProxyResolverFactory in the browser process to
serve as a proxy between the network service and the proxy resolver
service, starting the proxy resolver service as needed, but this
greatly simplifies what that class needs to do.

Bug: 754007
Reviewed-on: https://chromium-review.googlesource.com/753850
Commit-Queue: Matt Menke <mmenke@chromium.org>
Reviewed-by: Jay Civelli <jcivelli@chromium.org>
Reviewed-by: Tom Sepez <tsepez@chromium.org>
Reviewed-by: John Abd-El-Malek <jam@chromium.org>
Cr-Commit-Position: refs/heads/master@{#514303}