Add trace logs to profile the shared memory synchronization.
The progress of the shared memory synchronization is very important
on Castanets. We want to use Chrome trace to do that and time ticks
between processes need to be synchronized.

This CL still has a problem because the network delay affects timestamps.
We better check only total memories transferred without a duration.

Signed-off-by: Sunwoo Nam <sunny.nam@samsung.com>