Inform the browser of the shared memory created by the renderer.
i. In case of the command buffer, the renderer creates a shared memory
and then share it with the Browser. And both the browser and
renderer try to write data onto the shared memory and sync it.
At that point, the browser doesn't know which renderer(node) to
send sync data to. Thus, this CL adds a new broker message for
informing the browser of the shared memory created by the renderer.

ii. Fix code location to erase create_buffers_.

iii. Use broker channel to send the BUFFER_CREATED (non-sync msg) instead of
sync_channel_. This avoids receiving malformed message on browser side on
different test scenarios.

Signed-off-by: suyambu.rm <suyambu.rm@samsung.com>