Fix black tile problem, handle key events specific to tizen platform.
* Send InputHostMsg_DidHandleKeyEvent from renderer.

InputHostMsg_DidHandleKeyEvent IPC has to be sent to browser process
in case of tizen environment to handle key event.

Signed-off-by: suyambu.rm <suyambu.rm@samsung.com>

* Flush renderer commands before submitting frame.

There are chances when using tizen compositor (delegated)
GL commands could be executed out of order.
To fix this issue, add glFlush and glError before renderer
submits a frame to the browser.

Signed-off-by: suyambu.rm <suyambu.rm@samsung.com>