[Service Offloading] Set refresh interval to 60ms in Android.
Android media capturer doesn't capture if there is no new image.
During this time, chromium refreshes and sends packets once a second and
the bitrate is reduced.
When a new image begins to be captured again, there is a delay in increasing
the reduced bitrate.

By adjusting the refresh interval faster, bitrate is kept and
the delay problem is solved.

Signed-off-by: Sunwoo Nam <sunny.nam@samsung.com>