Fix DidFirstVisuallyNonEmptyPaint not received on browser
With DidFirstVisuallyNonEmptyPaint renderer informs the browser
that first layout has happened and helps browser to take a decision
as to when to show pixels on screen. Currently for all websites
DidFirstVisuallyNonEmptyPaint callback is not received on browser
side. Using Send method for sending IPC in this case helps and
similar changes present in tizen as well.

Signed-off-by: Uzair Jaleel <uzair.jaleel@samsung.com>