[Service Offloading] Allow getDisplayMedia in non-secure context.
In Service offloading, the android device gets contents for the webrtc server
from the client device. The android device needs to use http protocol to access
the client device, but getDisplayMedia does not work with http content.
So, this patch allows that getDisplayMedia works on the http content by setting
the document state to 'Secure'.

Signed-off-by: Sunwoo Nam <sunny.nam@samsung.com>